3.4.2 Integration tests with @SpringBootTest
[Web environment option]
MOCK: Creates a web application context with a mock Servlet container. This is the default option.
RANDOM_PORT: Creates a web application context with a Servlet container listening on a random port.
DEFINED_PORT: Creates a web application context with a Servlet container listening on the port defined through the server.port property.
NONE: Creates an application context without a Servlet container.